LayoutTests/imported/w3c/web-platform-tests/html/semantics/embedded-content/the-img-element/naturalWidth-naturalHeight-width-height.tentative-expected.txt



PASS raster image
PASS raster image (when not rendered)
PASS raster image with width/height attributes
PASS raster image with width/height attributes (when not rendered)
PASS non existent image, no natural dimensions
PASS non existent image, no natural dimensions (when not rendered)
PASS non existent image with width/height attributes, no natural dimensions
PASS non existent image with width/height attributes, no natural dimensions (when not rendered)
PASS SVG image, no natural dimensions
PASS SVG image, no natural dimensions (when not rendered)
FAIL SVG image with width/height attrs, no natural dimensions assert_equals: naturalWidth expected 300 but got 40
PASS SVG image with width/height attrs, no natural dimensions (when not rendered)
FAIL SVG image with width attr, no natural dimensions assert_equals: naturalWidth expected 300 but got 40
PASS SVG image with width attr, no natural dimensions (when not rendered)
FAIL SVG image with height attr, no natural dimensions assert_equals: naturalHeight expected 150 but got 10
PASS SVG image with height attr, no natural dimensions (when not rendered)
PASS SVG image, percengage natural dimensions
PASS SVG image, percengage natural dimensions (when not rendered)
PASS SVG image, negative percengage natural dimensions
PASS SVG image, negative percengage natural dimensions (when not rendered)
PASS SVG image, with natural width
FAIL SVG image, with natural width (when not rendered) assert_equals: naturalWidth when not rendered expected 60 but got 300
PASS SVG image, with natural height
FAIL SVG image, with natural height (when not rendered) assert_equals: naturalHeight when not rendered expected 60 but got 150
FAIL SVG image, with natural width of 0 assert_equals: naturalWidth expected 0 but got 300
FAIL SVG image, with natural width of 0 (when not rendered) assert_equals: naturalWidth when not rendered expected 0 but got 300
FAIL SVG image, with natural height of 0 assert_equals: naturalHeight expected 0 but got 150
FAIL SVG image, with natural height of 0 (when not rendered) assert_equals: naturalHeight when not rendered expected 0 but got 150
FAIL SVG image, with natural width being negative assert_equals: naturalWidth expected 0 but got 300
FAIL SVG image, with natural width being negative (when not rendered) assert_equals: naturalWidth when not rendered expected 0 but got 300
FAIL SVG image, with natural height being negative assert_equals: naturalHeight expected 0 but got 150
FAIL SVG image, with natural height being negative (when not rendered) assert_equals: naturalHeight when not rendered expected 0 but got 150
FAIL SVG image, no natural dimensions, and aspect ratio from viewBox assert_equals: naturalWidth expected 300 but got 720
FAIL SVG image, no natural dimensions, and aspect ratio from viewBox (when not rendered) assert_equals: naturalWidth when not rendered expected 300 but got 600
FAIL SVG image, percengage natural dimensions, and aspect ratio from viewBox assert_equals: naturalWidth expected 300 but got 720
FAIL SVG image, percengage natural dimensions, and aspect ratio from viewBox (when not rendered) assert_equals: naturalWidth when not rendered expected 300 but got 600
FAIL SVG image, negative percengage natural dimensions, and aspect ratio from viewBox assert_equals: naturalWidth expected 300 but got 720
FAIL SVG image, negative percengage natural dimensions, and aspect ratio from viewBox (when not rendered) assert_equals: naturalWidth when not rendered expected 300 but got 600
PASS SVG image, with natural width, and aspect ratio from viewBox
FAIL SVG image, with natural width, and aspect ratio from viewBox (when not rendered) assert_equals: naturalWidth when not rendered expected 60 but got 600
PASS SVG image, with natural height, and aspect ratio from viewBox
FAIL SVG image, with natural height, and aspect ratio from viewBox (when not rendered) assert_equals: naturalWidth when not rendered expected 180 but got 600
FAIL SVG image, with natural width of 0, and aspect ratio from viewBox assert_equals: naturalWidth expected 0 but got 720
FAIL SVG image, with natural width of 0, and aspect ratio from viewBox (when not rendered) assert_equals: naturalWidth when not rendered expected 0 but got 600
FAIL SVG image, with natural height of 0, and aspect ratio from viewBox assert_equals: naturalWidth expected 0 but got 720
FAIL SVG image, with natural height of 0, and aspect ratio from viewBox (when not rendered) assert_equals: naturalWidth when not rendered expected 0 but got 600
FAIL SVG image, with natural width being negative, and aspect ratio from viewBox assert_equals: naturalWidth expected 0 but got 720
FAIL SVG image, with natural width being negative, and aspect ratio from viewBox (when not rendered) assert_equals: naturalWidth when not rendered expected 0 but got 600
FAIL SVG image, with natural height being negative, and aspect ratio from viewBox assert_equals: naturalWidth expected 0 but got 720
FAIL SVG image, with natural height being negative, and aspect ratio from viewBox (when not rendered) assert_equals: naturalWidth when not rendered expected 0 but got 600
PASS SVG image, no natural dimensions, viewBox with 0 width/height
PASS SVG image, no natural dimensions, viewBox with 0 width/height (when not rendered)
PASS SVG image, no natural dimensions, viewBox with 0 width
PASS SVG image, no natural dimensions, viewBox with 0 width (when not rendered)
PASS SVG image, no natural dimensions, viewBox with 0 height
PASS SVG image, no natural dimensions, viewBox with 0 height (when not rendered)
PASS SVG image, with natural width, viewBox with 0 width/height
FAIL SVG image, with natural width, viewBox with 0 width/height (when not rendered) assert_equals: naturalWidth when not rendered expected 60 but got 300
PASS SVG image, with natural width, viewBox with 0 width
FAIL SVG image, with natural width, viewBox with 0 width (when not rendered) assert_equals: naturalWidth when not rendered expected 60 but got 300
PASS SVG image, with natural width, viewBox with 0 height
FAIL SVG image, with natural width, viewBox with 0 height (when not rendered) assert_equals: naturalWidth when not rendered expected 60 but got 300
PASS SVG image, with natural height, viewBox with 0 width/height
FAIL SVG image, with natural height, viewBox with 0 width/height (when not rendered) assert_equals: naturalHeight when not rendered expected 60 but got 150
PASS SVG image, with natural height, viewBox with 0 width
FAIL SVG image, with natural height, viewBox with 0 width (when not rendered) assert_equals: naturalHeight when not rendered expected 60 but got 150
PASS SVG image, with natural height, viewBox with 0 height
FAIL SVG image, with natural height, viewBox with 0 height (when not rendered) assert_equals: naturalHeight when not rendered expected 60 but got 150
PASS SVG image, with natural width and height
PASS SVG image, with natural width and height (when not rendered)
PASS SVG image, with natural width and height, and aspect ratio from viewBox
PASS SVG image, with natural width and height, and aspect ratio from viewBox (when not rendered)
FAIL SVG image, with natural width and height of 0, and aspect ratio from viewBox assert_equals: naturalWidth expected 0 but got 720
FAIL SVG image, with natural width and height of 0, and aspect ratio from viewBox (when not rendered) assert_equals: naturalWidth when not rendered expected 0 but got 300
FAIL SVG image, with natural width and height being negative, and aspect ratio from viewBox assert_equals: naturalWidth expected 0 but got 720
FAIL SVG image, with natural width and height being negative, and aspect ratio from viewBox (when not rendered) assert_equals: naturalWidth when not rendered expected 0 but got 600
PASS raster image (with srcset/1x)
PASS raster image (with srcset/1x) (when not rendered)
PASS raster image with width/height attributes (with srcset/1x)
PASS raster image with width/height attributes (with srcset/1x) (when not rendered)
PASS non existent image, no natural dimensions (with srcset/1x)
PASS non existent image, no natural dimensions (with srcset/1x) (when not rendered)
PASS non existent image with width/height attributes, no natural dimensions (with srcset/1x)
PASS non existent image with width/height attributes, no natural dimensions (with srcset/1x) (when not rendered)
PASS SVG image, no natural dimensions (with srcset/1x)
PASS SVG image, no natural dimensions (with srcset/1x) (when not rendered)
FAIL SVG image with width/height attrs, no natural dimensions (with srcset/1x) assert_equals: naturalWidth expected 300 but got 40
PASS SVG image with width/height attrs, no natural dimensions (with srcset/1x) (when not rendered)
FAIL SVG image with width attr, no natural dimensions (with srcset/1x) assert_equals: naturalWidth expected 300 but got 40
PASS SVG image with width attr, no natural dimensions (with srcset/1x) (when not rendered)
FAIL SVG image with height attr, no natural dimensions (with srcset/1x) assert_equals: naturalHeight expected 150 but got 10
PASS SVG image with height attr, no natural dimensions (with srcset/1x) (when not rendered)
PASS SVG image, percengage natural dimensions (with srcset/1x)
PASS SVG image, percengage natural dimensions (with srcset/1x) (when not rendered)
PASS SVG image, negative percengage natural dimensions (with srcset/1x)
PASS SVG image, negative percengage natural dimensions (with srcset/1x) (when not rendered)
PASS SVG image, with natural width (with srcset/1x)
FAIL SVG image, with natural width (with srcset/1x) (when not rendered) assert_equals: naturalWidth when not rendered expected 60 but got 300
PASS SVG image, with natural height (with srcset/1x)
FAIL SVG image, with natural height (with srcset/1x) (when not rendered) assert_equals: naturalHeight when not rendered expected 60 but got 150
FAIL SVG image, with natural width of 0 (with srcset/1x) assert_equals: naturalWidth expected 0 but got 300
FAIL SVG image, with natural width of 0 (with srcset/1x) (when not rendered) assert_equals: naturalWidth when not rendered expected 0 but got 300
FAIL SVG image, with natural height of 0 (with srcset/1x) assert_equals: naturalHeight expected 0 but got 150
FAIL SVG image, with natural height of 0 (with srcset/1x) (when not rendered) assert_equals: naturalHeight when not rendered expected 0 but got 150
FAIL SVG image, with natural width being negative (with srcset/1x) assert_equals: naturalWidth expected 0 but got 300
FAIL SVG image, with natural width being negative (with srcset/1x) (when not rendered) assert_equals: naturalWidth when not rendered expected 0 but got 300
FAIL SVG image, with natural height being negative (with srcset/1x) assert_equals: naturalHeight expected 0 but got 150
FAIL SVG image, with natural height being negative (with srcset/1x) (when not rendered) assert_equals: naturalHeight when not rendered expected 0 but got 150
FAIL SVG image, no natural dimensions, and aspect ratio from viewBox (with srcset/1x) assert_equals: naturalWidth expected 300 but got 720
FAIL SVG image, no natural dimensions, and aspect ratio from viewBox (with srcset/1x) (when not rendered) assert_equals: naturalWidth when not rendered expected 300 but got 600
FAIL SVG image, percengage natural dimensions, and aspect ratio from viewBox (with srcset/1x) assert_equals: naturalWidth expected 300 but got 720
FAIL SVG image, percengage natural dimensions, and aspect ratio from viewBox (with srcset/1x) (when not rendered) assert_equals: naturalWidth when not rendered expected 300 but got 600
FAIL SVG image, negative percengage natural dimensions, and aspect ratio from viewBox (with srcset/1x) assert_equals: naturalWidth expected 300 but got 720
FAIL SVG image, negative percengage natural dimensions, and aspect ratio from viewBox (with srcset/1x) (when not rendered) assert_equals: naturalWidth when not rendered expected 300 but got 600
PASS SVG image, with natural width, and aspect ratio from viewBox (with srcset/1x)
FAIL SVG image, with natural width, and aspect ratio from viewBox (with srcset/1x) (when not rendered) assert_equals: naturalWidth when not rendered expected 60 but got 600
PASS SVG image, with natural height, and aspect ratio from viewBox (with srcset/1x)
FAIL SVG image, with natural height, and aspect ratio from viewBox (with srcset/1x) (when not rendered) assert_equals: naturalWidth when not rendered expected 180 but got 600
FAIL SVG image, with natural width of 0, and aspect ratio from viewBox (with srcset/1x) assert_equals: naturalWidth expected 0 but got 720
FAIL SVG image, with natural width of 0, and aspect ratio from viewBox (with srcset/1x) (when not rendered) assert_equals: naturalWidth when not rendered expected 0 but got 600
FAIL SVG image, with natural height of 0, and aspect ratio from viewBox (with srcset/1x) assert_equals: naturalWidth expected 0 but got 720
FAIL SVG image, with natural height of 0, and aspect ratio from viewBox (with srcset/1x) (when not rendered) assert_equals: naturalWidth when not rendered expected 0 but got 600
FAIL SVG image, with natural width being negative, and aspect ratio from viewBox (with srcset/1x) assert_equals: naturalWidth expected 0 but got 720
FAIL SVG image, with natural width being negative, and aspect ratio from viewBox (with srcset/1x) (when not rendered) assert_equals: naturalWidth when not rendered expected 0 but got 600
FAIL SVG image, with natural height being negative, and aspect ratio from viewBox (with srcset/1x) assert_equals: naturalWidth expected 0 but got 720
FAIL SVG image, with natural height being negative, and aspect ratio from viewBox (with srcset/1x) (when not rendered) assert_equals: naturalWidth when not rendered expected 0 but got 600
PASS SVG image, no natural dimensions, viewBox with 0 width/height (with srcset/1x)
PASS SVG image, no natural dimensions, viewBox with 0 width/height (with srcset/1x) (when not rendered)
PASS SVG image, no natural dimensions, viewBox with 0 width (with srcset/1x)
PASS SVG image, no natural dimensions, viewBox with 0 width (with srcset/1x) (when not rendered)
PASS SVG image, no natural dimensions, viewBox with 0 height (with srcset/1x)
PASS SVG image, no natural dimensions, viewBox with 0 height (with srcset/1x) (when not rendered)
PASS SVG image, with natural width, viewBox with 0 width/height (with srcset/1x)
FAIL SVG image, with natural width, viewBox with 0 width/height (with srcset/1x) (when not rendered) assert_equals: naturalWidth when not rendered expected 60 but got 300
PASS SVG image, with natural width, viewBox with 0 width (with srcset/1x)
FAIL SVG image, with natural width, viewBox with 0 width (with srcset/1x) (when not rendered) assert_equals: naturalWidth when not rendered expected 60 but got 300
PASS SVG image, with natural width, viewBox with 0 height (with srcset/1x)
FAIL SVG image, with natural width, viewBox with 0 height (with srcset/1x) (when not rendered) assert_equals: naturalWidth when not rendered expected 60 but got 300
PASS SVG image, with natural height, viewBox with 0 width/height (with srcset/1x)
FAIL SVG image, with natural height, viewBox with 0 width/height (with srcset/1x) (when not rendered) assert_equals: naturalHeight when not rendered expected 60 but got 150
PASS SVG image, with natural height, viewBox with 0 width (with srcset/1x)
FAIL SVG image, with natural height, viewBox with 0 width (with srcset/1x) (when not rendered) assert_equals: naturalHeight when not rendered expected 60 but got 150
PASS SVG image, with natural height, viewBox with 0 height (with srcset/1x)
FAIL SVG image, with natural height, viewBox with 0 height (with srcset/1x) (when not rendered) assert_equals: naturalHeight when not rendered expected 60 but got 150
PASS SVG image, with natural width and height (with srcset/1x)
PASS SVG image, with natural width and height (with srcset/1x) (when not rendered)
PASS SVG image, with natural width and height, and aspect ratio from viewBox (with srcset/1x)
PASS SVG image, with natural width and height, and aspect ratio from viewBox (with srcset/1x) (when not rendered)
FAIL SVG image, with natural width and height of 0, and aspect ratio from viewBox (with srcset/1x) assert_equals: naturalWidth expected 0 but got 720
FAIL SVG image, with natural width and height of 0, and aspect ratio from viewBox (with srcset/1x) (when not rendered) assert_equals: naturalWidth when not rendered expected 0 but got 300
FAIL SVG image, with natural width and height being negative, and aspect ratio from viewBox (with srcset/1x) assert_equals: naturalWidth expected 0 but got 720
FAIL SVG image, with natural width and height being negative, and aspect ratio from viewBox (with srcset/1x) (when not rendered) assert_equals: naturalWidth when not rendered expected 0 but got 600
PASS raster image (with srcset/2x)
PASS raster image (with srcset/2x) (when not rendered)
PASS raster image with width/height attributes (with srcset/2x)
PASS raster image with width/height attributes (with srcset/2x) (when not rendered)
PASS non existent image, no natural dimensions (with srcset/2x)
PASS non existent image, no natural dimensions (with srcset/2x) (when not rendered)
PASS non existent image with width/height attributes, no natural dimensions (with srcset/2x)
PASS non existent image with width/height attributes, no natural dimensions (with srcset/2x) (when not rendered)
FAIL SVG image, no natural dimensions (with srcset/2x) assert_equals: width expected 300 but got 150
FAIL SVG image, no natural dimensions (with srcset/2x) (when not rendered) assert_equals: naturalWidth when not rendered expected 150 but got 300
FAIL SVG image with width/height attrs, no natural dimensions (with srcset/2x) assert_equals: naturalWidth expected 150 but got 40
FAIL SVG image with width/height attrs, no natural dimensions (with srcset/2x) (when not rendered) assert_equals: naturalWidth when not rendered expected 150 but got 300
FAIL SVG image with width attr, no natural dimensions (with srcset/2x) assert_equals: naturalWidth expected 150 but got 40
FAIL SVG image with width attr, no natural dimensions (with srcset/2x) (when not rendered) assert_equals: naturalWidth when not rendered expected 150 but got 300
FAIL SVG image with height attr, no natural dimensions (with srcset/2x) assert_equals: naturalHeight expected 75 but got 10
FAIL SVG image with height attr, no natural dimensions (with srcset/2x) (when not rendered) assert_equals: naturalWidth when not rendered expected 150 but got 300
FAIL SVG image, percengage natural dimensions (with srcset/2x) assert_equals: width expected 300 but got 150
FAIL SVG image, percengage natural dimensions (with srcset/2x) (when not rendered) assert_equals: naturalWidth when not rendered expected 150 but got 300
FAIL SVG image, negative percengage natural dimensions (with srcset/2x) assert_equals: width expected 300 but got 150
FAIL SVG image, negative percengage natural dimensions (with srcset/2x) (when not rendered) assert_equals: naturalWidth when not rendered expected 150 but got 300
FAIL SVG image, with natural width (with srcset/2x) assert_equals: height expected 150 but got 75
FAIL SVG image, with natural width (with srcset/2x) (when not rendered) assert_equals: naturalWidth when not rendered expected 30 but got 300
FAIL SVG image, with natural height (with srcset/2x) assert_equals: width expected 300 but got 150
FAIL SVG image, with natural height (with srcset/2x) (when not rendered) assert_equals: naturalWidth when not rendered expected 150 but got 300
FAIL SVG image, with natural width of 0 (with srcset/2x) assert_equals: naturalWidth expected 0 but got 150
FAIL SVG image, with natural width of 0 (with srcset/2x) (when not rendered) assert_equals: naturalWidth when not rendered expected 0 but got 300
FAIL SVG image, with natural height of 0 (with srcset/2x) assert_equals: naturalHeight expected 0 but got 75
FAIL SVG image, with natural height of 0 (with srcset/2x) (when not rendered) assert_equals: naturalWidth when not rendered expected 150 but got 300
FAIL SVG image, with natural width being negative (with srcset/2x) assert_equals: naturalWidth expected 0 but got 150
FAIL SVG image, with natural width being negative (with srcset/2x) (when not rendered) assert_equals: naturalWidth when not rendered expected 0 but got 300
FAIL SVG image, with natural height being negative (with srcset/2x) assert_equals: naturalHeight expected 0 but got 75
FAIL SVG image, with natural height being negative (with srcset/2x) (when not rendered) assert_equals: naturalWidth when not rendered expected 150 but got 300
FAIL SVG image, no natural dimensions, and aspect ratio from viewBox (with srcset/2x) assert_equals: naturalWidth expected 150 but got 720
FAIL SVG image, no natural dimensions, and aspect ratio from viewBox (with srcset/2x) (when not rendered) assert_equals: naturalWidth when not rendered expected 150 but got 600
FAIL SVG image, percengage natural dimensions, and aspect ratio from viewBox (with srcset/2x) assert_equals: naturalWidth expected 150 but got 720
FAIL SVG image, percengage natural dimensions, and aspect ratio from viewBox (with srcset/2x) (when not rendered) assert_equals: naturalWidth when not rendered expected 150 but got 600
FAIL SVG image, negative percengage natural dimensions, and aspect ratio from viewBox (with srcset/2x) assert_equals: naturalWidth expected 150 but got 720
FAIL SVG image, negative percengage natural dimensions, and aspect ratio from viewBox (with srcset/2x) (when not rendered) assert_equals: naturalWidth when not rendered expected 150 but got 600
PASS SVG image, with natural width, and aspect ratio from viewBox (with srcset/2x)
FAIL SVG image, with natural width, and aspect ratio from viewBox (with srcset/2x) (when not rendered) assert_equals: naturalWidth when not rendered expected 30 but got 600
PASS SVG image, with natural height, and aspect ratio from viewBox (with srcset/2x)
FAIL SVG image, with natural height, and aspect ratio from viewBox (with srcset/2x) (when not rendered) assert_equals: naturalWidth when not rendered expected 90 but got 600
FAIL SVG image, with natural width of 0, and aspect ratio from viewBox (with srcset/2x) assert_equals: naturalWidth expected 0 but got 720
FAIL SVG image, with natural width of 0, and aspect ratio from viewBox (with srcset/2x) (when not rendered) assert_equals: naturalWidth when not rendered expected 0 but got 600
FAIL SVG image, with natural height of 0, and aspect ratio from viewBox (with srcset/2x) assert_equals: naturalWidth expected 0 but got 720
FAIL SVG image, with natural height of 0, and aspect ratio from viewBox (with srcset/2x) (when not rendered) assert_equals: naturalWidth when not rendered expected 0 but got 600
FAIL SVG image, with natural width being negative, and aspect ratio from viewBox (with srcset/2x) assert_equals: naturalWidth expected 0 but got 720
FAIL SVG image, with natural width being negative, and aspect ratio from viewBox (with srcset/2x) (when not rendered) assert_equals: naturalWidth when not rendered expected 0 but got 600
FAIL SVG image, with natural height being negative, and aspect ratio from viewBox (with srcset/2x) assert_equals: naturalWidth expected 0 but got 720
FAIL SVG image, with natural height being negative, and aspect ratio from viewBox (with srcset/2x) (when not rendered) assert_equals: naturalWidth when not rendered expected 0 but got 600
FAIL SVG image, no natural dimensions, viewBox with 0 width/height (with srcset/2x) assert_equals: width expected 300 but got 150
FAIL SVG image, no natural dimensions, viewBox with 0 width/height (with srcset/2x) (when not rendered) assert_equals: naturalWidth when not rendered expected 150 but got 300
FAIL SVG image, no natural dimensions, viewBox with 0 width (with srcset/2x) assert_equals: width expected 300 but got 150
FAIL SVG image, no natural dimensions, viewBox with 0 width (with srcset/2x) (when not rendered) assert_equals: naturalWidth when not rendered expected 150 but got 300
FAIL SVG image, no natural dimensions, viewBox with 0 height (with srcset/2x) assert_equals: width expected 300 but got 150
FAIL SVG image, no natural dimensions, viewBox with 0 height (with srcset/2x) (when not rendered) assert_equals: naturalWidth when not rendered expected 150 but got 300
FAIL SVG image, with natural width, viewBox with 0 width/height (with srcset/2x) assert_equals: height expected 150 but got 75
FAIL SVG image, with natural width, viewBox with 0 width/height (with srcset/2x) (when not rendered) assert_equals: naturalWidth when not rendered expected 30 but got 300
FAIL SVG image, with natural width, viewBox with 0 width (with srcset/2x) assert_equals: height expected 150 but got 75
FAIL SVG image, with natural width, viewBox with 0 width (with srcset/2x) (when not rendered) assert_equals: naturalWidth when not rendered expected 30 but got 300
FAIL SVG image, with natural width, viewBox with 0 height (with srcset/2x) assert_equals: height expected 150 but got 75
FAIL SVG image, with natural width, viewBox with 0 height (with srcset/2x) (when not rendered) assert_equals: naturalWidth when not rendered expected 30 but got 300
FAIL SVG image, with natural height, viewBox with 0 width/height (with srcset/2x) assert_equals: width expected 300 but got 150
FAIL SVG image, with natural height, viewBox with 0 width/height (with srcset/2x) (when not rendered) assert_equals: naturalWidth when not rendered expected 150 but got 300
FAIL SVG image, with natural height, viewBox with 0 width (with srcset/2x) assert_equals: width expected 300 but got 150
FAIL SVG image, with natural height, viewBox with 0 width (with srcset/2x) (when not rendered) assert_equals: naturalWidth when not rendered expected 150 but got 300
FAIL SVG image, with natural height, viewBox with 0 height (with srcset/2x) assert_equals: width expected 300 but got 150
FAIL SVG image, with natural height, viewBox with 0 height (with srcset/2x) (when not rendered) assert_equals: naturalWidth when not rendered expected 150 but got 300
PASS SVG image, with natural width and height (with srcset/2x)
FAIL SVG image, with natural width and height (with srcset/2x) (when not rendered) assert_equals: naturalWidth when not rendered expected 30 but got 60
PASS SVG image, with natural width and height, and aspect ratio from viewBox (with srcset/2x)
FAIL SVG image, with natural width and height, and aspect ratio from viewBox (with srcset/2x) (when not rendered) assert_equals: naturalWidth when not rendered expected 30 but got 60
FAIL SVG image, with natural width and height of 0, and aspect ratio from viewBox (with srcset/2x) assert_equals: naturalWidth expected 0 but got 720
FAIL SVG image, with natural width and height of 0, and aspect ratio from viewBox (with srcset/2x) (when not rendered) assert_equals: naturalWidth when not rendered expected 0 but got 300
FAIL SVG image, with natural width and height being negative, and aspect ratio from viewBox (with srcset/2x) assert_equals: naturalWidth expected 0 but got 720
FAIL SVG image, with natural width and height being negative, and aspect ratio from viewBox (with srcset/2x) (when not rendered) assert_equals: naturalWidth when not rendered expected 0 but got 600
